+++ src/de/lmu/ifi/dbs/elki/math/linearalgebra/LUDecomposition.java
@@ -189,7 +189,7 @@ public class LUDecomposition implements java.io.Serializable {
    public double[] getDoublePivot () {
       double[] vals = new double[m];
       for (int i = 0; i < m; i++) {
-         vals[i] = (double) piv[i];
+         vals[i] = piv[i];
       }
       return vals;
    }
@@ -203,7 +203,7 @@ public class LUDecomposition implements java.io.Serializable {
       if (m != n) {
          throw new IllegalArgumentException("Matrix must be square.");
       }
-      double d = (double) pivsign;
+      double d = pivsign;
       for (int j = 0; j < n; j++) {
          d *= LU[j][j];
       }
